WIP: async table api

+++ samza-core/src/main/java/org/apache/samza/table/remote/RemoteReadWriteTable.java
@@ -23,17 +23,15 @@ import java.util.List;
 
 import org.apache.samza.SamzaException;
 import org.apache.samza.container.SamzaContainerContext;
-import org.apache.samza.metrics.Timer;
 import org.apache.samza.storage.kv.Entry;
 import org.apache.samza.table.ReadWriteTable;
+import org.apache.samza.table.TableOpCallback;
 import org.apache.samza.table.utils.DefaultTableWriteMetrics;
 import org.apache.samza.table.utils.TableMetricsUtil;
 import org.apache.samza.task.TaskContext;
-import org.apache.samza.util.RateLimiter;
 
+import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
-
-import static org.apache.samza.table.remote.RemoteTableDescriptor.RL_WRITE_TAG;
 
 
 /**
@@ -43,22 +41,19 @@ import static org.apache.samza.table.remote.RemoteTableDescriptor.RL_WRITE_TAG;
  * @param <V> the type of the value in this table
  */
 public class RemoteReadWriteTable<K, V> extends RemoteReadableTable<K, V> implements ReadWriteTable<K, V> {
+  private final TableWriteFunction<K, V> writeFn;
 
-  protected final TableWriteFunction<K, V> writeFn;
-  protected final CreditFunction<K, V> writeCreditFn;
-  protected final boolean rateLimitWrites;
+  private DefaultTableWriteMetrics writeMetrics;
 
-  protected DefaultTableWriteMetrics writeMetrics;
-  protected Timer putThrottleNs; // use single timer for all write operations
+  @VisibleForTesting
+  Throttler writeThrottler;
 
   public RemoteReadWriteTable(String tableId, TableReadFunction readFn, TableWriteFunction writeFn,
-      RateLimiter ratelimiter, CreditFunction<K, V> readCreditFn, CreditFunction<K, V> writeCreditFn) {
-    super(tableId, readFn, ratelimiter, readCreditFn);
+      Throttler<K, V> readThrottler, Throttler<K, V> writeThrottler, RequestManager<K, V> requestManager) {
+    super(tableId, readFn, readThrottler, requestManager);
     Preconditions.checkNotNull(writeFn, "null write function");
     this.writeFn = writeFn;
-    this.writeCreditFn = writeCreditFn;
-    this.rateLimitWrites = rateLimiter != null && rateLimiter.getSupportedTags().contains(RL_WRITE_TAG);
-    logger.info("Rate limiting is {} for remote write operations", rateLimitWrites ? "enabled" : "disabled");
+    this.writeThrottler = writeThrottler;
   }
 
   /**
@@ -69,7 +64,7 @@ public class RemoteReadWriteTable<K, V> extends RemoteReadableTable<K, V> implem
     super.init(containerContext, taskContext);
     writeMetrics = new DefaultTableWriteMetrics(containerContext, taskContext, this, tableId);
     TableMetricsUtil tableMetricsUtil = new TableMetricsUtil(containerContext, taskContext, this, tableId);
-    putThrottleNs = tableMetricsUtil.newTimer("put-throttle-ns");
+    writeThrottler.setTimerMetric(tableMetricsUtil.newTimer("put-throttleKeys-ns"));
   }
 
   /**
@@ -77,7 +72,7 @@ public class RemoteReadWriteTable<K, V> extends RemoteReadableTable<K, V> implem
    */
   @Override
   public void put(K key, V value) {
-
+    Preconditions.checkNotNull(key);
     if (value == null) {
       delete(key);
       return;
@@ -85,11 +80,8 @@ public class RemoteReadWriteTable<K, V> extends RemoteReadableTable<K, V> implem
 
     try {
       writeMetrics.numPuts.inc();
-      if (rateLimitWrites) {
-        throttle(key, value, RL_WRITE_TAG, writeCreditFn, putThrottleNs);
-      }
       long startNs = System.nanoTime();
-      writeFn.put(key, value);
+      requestManager.execute(key, value, () -> writeFn.put(key, value), writeThrottler);
       writeMetrics.putNs.update(System.nanoTime() - startNs);
     } catch (Exception e) {
       String errMsg = String.format("Failed to put a record, key=%s, value=%s", key, value);
@@ -102,11 +94,41 @@ public class RemoteReadWriteTable<K, V> extends RemoteReadableTable<K, V> implem
    * {@inheritDoc}
    */
   @Override
+  public void put(K key, V value, TableOpCallback callback) {
+    Preconditions.checkNotNull(key);
+    if (value == null) {
+      delete(key, callback);
+      return;
+    }
+
+    writeMetrics.numPuts.inc();
+
+    requestManager.executeAsync(key, value, () -> {
+      final long startNs = System.nanoTime();
+      writeFn.put(key, value, requestManager.decorate((dummy, error) -> {
+        writeMetrics.putNs.update(System.nanoTime() - startNs);
+        if (error != null) {
+          logger.error(String.format("Failed to put a record, key=%s, value=%s", key, value), error);
+        }
+        callbackHelper.invoke(callback, dummy, error, writeMetrics.putCallbackNs);
+      }));
+    }, writeThrottler);
+  }
+
+  /**
+   * {@inheritDoc}
+   */
+  @Override
   public void putAll(List<Entry<K, V>> entries) {
+    Preconditions.checkNotNull(entries);
+    if (entries.isEmpty()) {
+      return;
+    }
+
     try {
       writeMetrics.numPutAlls.inc();
       long startNs = System.nanoTime();
-      writeFn.putAll(entries);
+      requestManager.executeEntries(entries, () -> writeFn.putAll(entries), writeThrottler);
       writeMetrics.putAllNs.update(System.nanoTime() - startNs);
     } catch (Exception e) {
       String errMsg = String.format("Failed to put records: %s", entries);
@@ -115,18 +137,37 @@ public class RemoteReadWriteTable<K, V> extends RemoteReadableTable<K, V> implem
     }
   }
 
+  @Override
+  public void putAll(List<Entry<K, V>> entries, TableOpCallback callback) {
+    Preconditions.checkNotNull(entries);
+    if (entries.isEmpty()) {
+      callbackHelper.invoke(callback, null, null, writeMetrics.putCallbackNs);
+      return;
+    }
+
+    writeMetrics.numPutAlls.inc();
+    final long startNs = System.nanoTime();
+
+    requestManager.executeEntriesAsync(entries, () ->
+      writeFn.putAll(entries, requestManager.decorate((dummy, error) -> {
+        writeMetrics.putAllNs.update(System.nanoTime() - startNs);
+        if (error != null) {
+          logger.error(String.format("Failed to put records: %s", entries), error);
+        }
+        callbackHelper.invoke(callback, dummy, error, writeMetrics.putCallbackNs);
+      })), writeThrottler);
+  }
+
   /**
    * {@inheritDoc}
    */
   @Override
   public void delete(K key) {
+    Preconditions.checkNotNull(key);
     try {
       writeMetrics.numDeletes.inc();
-      if (rateLimitWrites) {
-        throttle(key, null, RL_WRITE_TAG, writeCreditFn, putThrottleNs);
-      }
       long startNs = System.nanoTime();
-      writeFn.delete(key);
+      requestManager.execute(key, null, () -> writeFn.delete(key), writeThrottler); // FIXME
       writeMetrics.deleteNs.update(System.nanoTime() - startNs);
     } catch (Exception e) {
       String errMsg = String.format("Failed to delete a record, key=%s", key);
@@ -135,21 +176,63 @@ public class RemoteReadWriteTable<K, V> extends RemoteReadableTable<K, V> implem
     }
   }
 
+  @Override
+  public void delete(K key, TableOpCallback callback) {
+    Preconditions.checkNotNull(key);
+    Preconditions.checkNotNull(callback);
+    writeMetrics.numDeletes.inc();
+    final long startNs = System.nanoTime();
+    requestManager.executeAsync(key, () -> {
+      writeFn.delete(key, requestManager.decorate((dummy, error) -> {
+        writeMetrics.deleteNs.update(System.nanoTime() - startNs);
+        if (error != null) {
+          logger.error(String.format("Failed to delete a record, key=%s", key), error);
+        }
+        callbackHelper.invoke(callback, dummy, error, writeMetrics.deleteCallbackNs);
+      }));
+    }, writeThrottler);
+  }
+
   /**
    * {@inheritDoc}
    */
   @Override
   public void deleteAll(List<K> keys) {
+    Preconditions.checkNotNull(keys);
+    if (keys.isEmpty()) {
+      return;
+    }
     try {
       writeMetrics.numDeleteAlls.inc();
-      writeFn.deleteAll(keys);
       long startNs = System.nanoTime();
+      requestManager.execute(keys, () -> { writeFn.deleteAll(keys); return null; }, writeThrottler); // FIXME
       writeMetrics.deleteAllNs.update(System.nanoTime() - startNs);
     } catch (Exception e) {
       String errMsg = String.format("Failed to delete records, keys=%s", keys);
       logger.error(errMsg, e);
       throw new SamzaException(errMsg, e);
     }
+  }
+  @Override
+  public void deleteAll(List<K> keys, TableOpCallback callback) {
+    Preconditions.checkNotNull(keys);
+    Preconditions.checkNotNull(callback);
+    if (keys.isEmpty()) {
+      callbackHelper.invoke(callback, null, null, writeMetrics.deleteCallbackNs);
+      return;
+    }
+    writeMetrics.numDeleteAlls.inc();
+    final long startNs = System.nanoTime();
+
+    requestManager.executeAsync(keys, () -> {
+      writeFn.deleteAll(keys, requestManager.decorate((dummy, error) -> {
+        writeMetrics.deleteAllNs.update(System.nanoTime() - startNs);
+        if (error != null) {
+          logger.error(String.format("Failed to delete records, keys=%s", keys), error);
+        }
+        callbackHelper.invoke(callback, dummy, error, writeMetrics.deleteCallbackNs);
+      }));
+    }, writeThrottler);
   }
 
   /**
@@ -159,9 +242,6 @@ public class RemoteReadWriteTable<K, V> extends RemoteReadableTable<K, V> implem
   public void flush() {
     try {
       writeMetrics.numFlushes.inc();
-      if (rateLimitWrites) {
-        throttle(null, null, RL_WRITE_TAG, writeCreditFn, putThrottleNs);
-      }
       long startNs = System.nanoTime();
       writeFn.flush();
       writeMetrics.flushNs.update(System.nanoTime() - startNs);
